+++ module-core/src/main/java/com/example/db/user/repository/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
   // 뷰
-  Optional<User> findByUserIdxAndIsDelete(Integer userIdx, String isDelete);
+  Optional<User> findByUserIdxAndIsDelete(Long userIdx, String isDelete);
 
   // 로그인
   Optional<User> findByIdAndIsDelete(String id, String isDelete);
